Apply quickcancel
Cancel slock by moving the mouse within a certain time-period after
slock started. The time-period can be defined in seconds with the
setting *timetocancel* in the config.h. This is useful if you forgot to
disable `xautolock` during an activity that requires no input (e.g.
reading text, watching video).